refactor(infra): fold aether-hub-protocol into codec + hub (ADR-0072)
Splits the retired hub-protocol crate along the mechanism / vocabulary
seam: generic length-prefixed postcard framing helpers move into
aether-codec::frame; hub channel frame types (EngineToHub, HubToEngine,
Hello, MailFrame, etc.) move into aether-hub::wire next to the runtime
that speaks them. Workspace member retired; three vestigial Cargo.toml
deps in aether-component / aether-scenario / aether-substrate-headless
pruned. Reverses the ADR-0069 reservation that kept hub-protocol as a
separate crate for a hypothetical sibling transport — no consumer of
the wire crate ever wanted it without also pulling the hub runtime.
Wire format unchanged; tests rewired to import from the new locations.
