+++ crates/aether-codec/Cargo.toml
@@ -24,6 +24,8 @@ path = "src/lib.rs"
 
 [dependencies]
 aether-data = { path = "../aether-data" }
+postcard = { version = "1.1", features = ["alloc"] }
+serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 
 [dev-dependencies]
@@ -36,5 +38,3 @@ serde_json = "1"
 aether-kinds = { path = "../aether-kinds" }
 aether-data = { path = "../aether-data", features = ["derive"] }
 bytemuck = { version = "1.19", features = ["derive"] }
-postcard = { version = "1.1", features = ["alloc"] }
-serde = { version = "1", features = ["derive"] }
