+++ crates/aether-codec/src/lib.rs
@@ -1,22 +1,30 @@
-//! Schema-driven encode + decode of agent-supplied params against
-//! `aether_data::SchemaType` descriptors. Pure functions —
-//! no hub state, no async — so callers outside the supervisor
-//! (scenario runner, future tooling) can use the same JSON ↔
-//! wire-bytes path
-//! the hub uses for `mcp__aether-hub__send_mail` and
-//! `mcp__aether-hub__receive_mail`.
+//! Byte encoding toolkit. Two layers, both pure functions / no async:
 //!
-//! Two wire shapes, picked per descriptor (ADR-0019, ADR-0020):
+//! - **Schema-driven encode/decode** ([`encode_schema`] / [`decode_schema`]):
+//!   walks an `aether_data::SchemaType` descriptor to encode agent-
+//!   supplied JSON params into the wire bytes the substrate's decode
+//!   path is happy with (and back out). Two wire shapes, picked per
+//!   descriptor (ADR-0019, ADR-0020):
 //!
-//! 1. Cast-shaped (`Struct { repr_c: true }` and the recursive tree
-//!    under it): `#[repr(C)]` byte layout. Decode is `bytemuck::cast`
-//!    on the substrate side; encode walks the schema and writes the
-//!    same layout.
-//! 2. Postcard (everything else): postcard 1.x wire format, written
-//!    and read directly to match the format byte-for-byte.
+//!   1. Cast-shaped (`Struct { repr_c: true }` and the recursive tree
+//!      under it): `#[repr(C)]` byte layout. Decode is `bytemuck::cast`
+//!      on the substrate side; encode walks the schema and writes the
+//!      same layout.
+//!   2. Postcard (everything else): postcard 1.x wire format, written
+//!      and read directly to match the format byte-for-byte.
+//!
+//! - **Stream framing** ([`frame`]): length-prefixed postcard for
+//!   serde-derived enum types. The hub channel (`aether_hub::wire`)
+//!   is the first consumer; ADR-0072 placed framing here because the
+//!   helpers are codec-shaped and generic over `<T: Serialize>`.
+//!
+//! Future formats (msgpack, protobuf, save-format adapters) land as
+//! sibling modules. Future framing variants subdivide [`frame`] under
+//! `frame::postcard` / `frame::protobuf`.
 
 mod decode;
 mod encode;
+pub mod frame;
 
 pub use decode::{DecodeError, decode_schema};
 pub use encode::{EncodeError, encode_schema};
