--- crates/aether-hub-protocol/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "aether-hub-protocol"
-version.workspace = true
-edition.workspace = true
-
-# ADR-0069 split out the universal data layer into `aether-data`. This
-# crate now carries only the engine ↔ hub channel wire (frames +
-# framing helpers), which is unambiguously host-side std code — the
-# `cfg_attr(not(feature = "std"), no_std)` gymnastics retired with
-# the schema-vocab move.
-[dependencies]
-aether-data = { path = "../aether-data" }
-postcard = { version = "1.1", features = ["alloc"] }
-serde = { version = "1", features = ["derive"] }
-uuid = { version = "1", features = ["serde"] }
