+++ crates/aether-substrate-desktop/Cargo.toml
@@ -18,7 +18,6 @@ path = "src/main.rs"
 [dependencies]
 aether-substrate-core = { path = "../aether-substrate-core", features = ["render", "audio"] }
 aether-hub = { path = "../aether-hub" }
-aether-hub-protocol = { path = "../aether-hub-protocol" }
 aether-data = { path = "../aether-data" }
 aether-kinds = { path = "../aether-kinds" }
 bytemuck = "1.19"
@@ -33,4 +32,5 @@ wgpu = "29.0.1"
 winit = "0.30.13"
 
 [dev-dependencies]
+aether-codec = { path = "../aether-codec" }
 wat = "1"
