+++ crates/aether-substrate-headless/Cargo.toml
@@ -12,7 +12,6 @@ edition.workspace = true
 [dependencies]
 aether-substrate-core = { path = "../aether-substrate-core" }
 aether-hub = { path = "../aether-hub" }
-aether-hub-protocol = { path = "../aether-hub-protocol" }
 aether-data = { path = "../aether-data" }
 aether-kinds = { path = "../aether-kinds" }
 tracing = "0.1"
