+++ crates/aether-substrate-test-bench/Cargo.toml
@@ -23,7 +23,6 @@ path = "src/lib.rs"
 [dependencies]
 aether-substrate-core = { path = "../aether-substrate-core", features = ["render"] }
 aether-hub = { path = "../aether-hub" }
-aether-hub-protocol = { path = "../aether-hub-protocol" }
 aether-data = { path = "../aether-data" }
 aether-kinds = { path = "../aether-kinds" }
 bytemuck = "1.19"
